issue on old methods

+++ Src/Models/ClassResponseType.cs
@@ -1,6 +1,6 @@
 ﻿namespace Xcsb.Models;
 
-public enum ClassResponseType : byte
+public enum ClassResponseType : ushort
 {
     InputOutput = 1,
     InputOnly = 2
